test: polish graphql custom predicate tests
